+++ src/experiment_prototype/scan_classes/sequences.py
@@ -25,9 +25,9 @@ import numpy as np
 import structlog
 
 # local
-from experiment_prototype.experiment_utils.sample_building import get_samples, get_phase_shift
 from experiment_prototype.scan_classes.scan_class_base import ScanClassBase
 from experiment_prototype.experiment_exception import ExperimentException
+from utils.signals import get_samples, get_phase_shift
 
 # Obtain the module name that imported this log_config
 caller = Path(inspect.stack()[-1].filename)
